remove turso db/group update CLI methods
