+++ internal/cmd/group_flag.go
@@ -19,12 +19,6 @@ import (
 	"github.com/tursodatabase/turso-cli/internal/turso"
 )
 
-var groupBoolFlag bool
-
-func addGroupBoolFlag(cmd *cobra.Command, description string) {
-	cmd.Flags().BoolVar(&groupBoolFlag, "group", false, description)
-}
-
 var groupFlag string
 
 func addGroupFlag(cmd *cobra.Command) {
